+++ knowledge/pages/about/CHANGELOG.md
@@ -15,6 +15,10 @@ tags: about
 - PHP functions to match write_log_*().
 - PHP errors are now logged to the Cloudy logfile.
 
+### Removed
+
+- Composer vendor and composer.lock are no longer shipped.
+
 ## [1.7.1] - 2024-05-08
 
 ### Fixed